set up tests

+++ microsim-core/src/main/java/microsim/engine/SimulationEngine.java
@@ -438,7 +438,7 @@ public class SimulationEngine extends Thread {
 	 *            The new random seed.
 	 */
 	public void setRandomSeed(long newSeed) {
-		rnd.setSeed((int) newSeed);
+		rnd.setSeed(newSeed);
 		randomSeed = newSeed;
 	}
